Improved `start_install.sh` and `finish_install.sh`
Added ternary operators to many lines of code as a way to print a
helpful message as to why the command failed, and then to exit to
prevent further commands from being executed after one fails. Moved
around and optimized much of the code to reduce the size of the scripts.
Removed some functions and improved others.