Added a testing directory + small improvements
Added a new testing directory in which you can can use the Makefile to
run `make test` to create a fresh Virtual Machine with QEMU for testing
on. Fixed an error with setting the user password in `finish_install.sh`,
and removed the enabling of the snapper-timeline & snapper-cleanup
services as they never worked well and the user should choose whether
they want to enable them. Added a new prompt that uses fzf to let the
user choose there timezone.